aggregator/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
kafka-python==2.0.2
pydantic==1.10.11
python-dotenv==1.0.0
collections-extended==0.1.2
